Immediately send pending changes when backgrounding app.
